refactor: Consistent css folders
